+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.1
+
+- add non minified version
+
 ## 0.1.0
 
 - Fix "map has not control" when RMap cleanup is called before useControl cleanup.